refactor(compiler): one lowering for `for` and `if` in both source positions
Raku control constructs can be written as a statement (`for @xs { ... }`) or
as an expression that produces a value (`my @r = do for @xs { ... }`). mutsu
compiled the two positions through two independent code paths -- the
`Stmt::For` / `Stmt::If` arms of `compile_stmt` and the `compile_do_*` family
in `helpers_do_expr.rs` -- so every control-flow semantic had to be written
twice. The clearest evidence was `ForLoopSpec`, the 30-field struct that
carries a compiled `for` loop to `OpCode::ForLoop`: it was built by hand in
two places with nothing enforcing that the literals agreed.

Add two modules holding one lowering each, and reduce the position-specific
callers to thin wrappers:

- `compiler/control_for.rs` -- `compile_for_construct(ForParts)`, where
  `ForParts::collect` is the only thing the two positions differ in.
  `ForLoopSpec` now has exactly one construction site in the compiler.
- `compiler/control_if.rs` -- `compile_if_construct(..., IfPosition)`, sharing
  the condition handling, the duplicated-condition value that feeds `@_` and
  placeholder binding, the jump/patch structure, the pointy-topic scope and
  the `elsif` recursion.

`Stmt::For` keeps its statement-level source desugars (`for @a[*]`,
element-source writeback) and `Stmt::If` its statement-level pre-checks
(heredoc scope, constant-condition folding): both rewrite the whole construct
and re-enter `compile_stmt`. The `while` / C-style `loop` / `lazy for`
expression forms were already thin `gather`-desugaring wrappers over the
statement path and are unchanged.

Merging exposed six value-position divergences, all now fixed: loop phasers
were dropped entirely (`FIRST`/`LAST` never fired in a value-position `for`);
a `.reverse`d rw source wrote back mirrored (`TagContainerRef` instead of
`TagContainerRefReversed`); a typed loop parameter did not autothread
junctions; the loop parameter was not recorded in `for_loop_param_syms`;
multi-params were neither marked read-only nor recorded in `param_bind_names`;
and `uses_block_magic` never reached the value path, so `&?BLOCK` got no block
callable. On the `if` side the value path used `compile_expr` rather than
`compile_condition_expr`, so a bare regex condition was an always-true literal
instead of a smartmatch against the topic.

`helpers_do_expr.rs` 666 -> 398 lines, `stmt.rs` 5191 -> 4740.

Two pre-existing value-position bugs surfaced during the audit and are filed
rather than fixed here, since neither is part of the lowering: a `NEXT` phaser
clobbering the collected value, and a nested bare block's `state` not
restarting per iteration.

Pin: t/control-construct-value-position.t.
